fix NT matches not appearing in MyMonitor
This was primarily failing because CHPP matches.xml is no longer supposed to
be used for NT/U21 matches. Technically NT matches are now HTO SourceSystem
and when accessed via matches.xml the MatchType will be an HTO type. The
existing code has been hacked to work around this, but we should really be
using nationalteammatches.xml for this.